backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
pydantic==2.10.3
python-multipart==0.0.12
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.2.3
openpyxl==3.1.5
